+++ CHANGELOG.md
@@ -1,11 +1,21 @@
 # Change Log
 
-All notable changes to the "vscode-codeigniter4-spark" extension will be documented in this file.
+All notable changes to the "Codeigniter 4 Spark for VSCode" extension will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.1.0] - 2023-08-28
+
+### Added
+
+- Add Resource Command
+
+### Fixed
+
+- Fix typos
 
 ## [1.0.1] - 2023-08-27
 
@@ -17,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
-[unreleased]: https://github.com/ManuelGil/vscode-codeigniter4-spark/compare/v1.0.1...HEAD
-[1.0.1]: https://github.com/ManuelGil/vscode-codeigniter4-snippets/compare/v1.0.0...v1.0.1
+[unreleased]: https://github.com/ManuelGil/vscode-codeigniter4-spark/compare/v1.1.0...HEAD
+[1.1.0]: https://github.com/ManuelGil/vscode-codeigniter4-spark/compare/v1.0.1...v1.1.0
+[1.0.1]: https://github.com/ManuelGil/vscode-codeigniter4-spark/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/ManuelGil/vscode-codeigniter4-spark/releases/tag/v1.0.0